update_valoracion

+++ PruebaDBP/Models/Pelicula.cs
@@ -14,14 +14,14 @@ namespace PruebaDBP.Models
         [Required(ErrorMessage = "El campo fecha de esteno es obligatorio")]
         public string? FechaEstreno { get; set; }
         [Required(ErrorMessage = "El campo valoracion es obligatorio")]
-        public int? Valoracion { get; set; }
+        public decimal? Valoracion { get; set; }
         [Required(ErrorMessage = "El campo sumilla es obligatorio")]
         public string? Sumilla { get; set; }
         [Required(ErrorMessage = "El campo UrlFoto es obligatorio")]
         public string? UrlFoto { get; set; }
 
 
-        public Pelicula(int? idTmdb, int? idIdioma, string? nomPelicula, string? fechaEstreno, int? valoracion, string? sumilla, string? urlFoto)
+        public Pelicula(int? idTmdb, int? idIdioma, string? nomPelicula, string? fechaEstreno, decimal? valoracion, string? sumilla, string? urlFoto)
         {
             IdTmdb = idTmdb;
             IdIdioma = idIdioma;
