+++ PruebaDBP/Models/PeliculaEstanterium.cs
@@ -31,7 +31,7 @@ namespace PruebaDBP.Models
         public string? FechaAgregacion { get; set; }
         public int? DuracionMin { get; set; }
         public string? Sumilla { get; set; }
-        public int? Valoracion { get; set; }
+        public decimal? Valoracion { get; set; }
         public List<Director> directores{get;set;}
         public ValoracionUsuario valoracion;
